Rename accordion-classes to accordion-summary-classes
