--- packages/mui-codemod/src/deprecations/accordion-classes/index.js
@@ -1,1 +1,0 @@
-export { default } from './accordion-classes';
