+++ packages/mui-codemod/src/deprecations/accordion-summary-classes/index.js
@@ -1,0 +1,1 @@
+export { default } from './accordion-summary-classes';
