+++ packages/mui-codemod/src/deprecations/accordion-summary-classes/test-cases/expected.js
